[WindowDecor] Don't create maximize menu if mid-drag
This CL skips creating a maximize menu (triggered by on long click
listener) if we are mid-drag. If the maximize menu is already active
when we start dragging, we will also close the maximize menu.

Bug: 364382694
Test: Manual - drag the task by long pressing on the maximize button and
verify maximize menu does not show. Long press on maximize button to
show maximize menu, then drag the task window and verify the menu is
closed.
Flag: EXEMPT BUGFIX

Change-Id: I2d188659c8275a0d00edf5707e5b877350566f9f