Pass Configuration Parameters to Handoff
When calling Activity.setHandoffEnabled(), allow an optional configuration structure to be passed in. This allows the Activity to declare if it is capable of handing off to devices without the current package installed.

Under-the-hood, persist this structure as a data storage mechanism on ActivityRecord, replacing the two booleans.

Flag: android.companion.task_continuity
Bug: 400970610
Test: Added Unit Tests


Change-Id: I4f5f2bbda4853fa9b14c461930ba73173818e47f