Handle changes to App Lock enabled state in AppLockLocalService
Register a PackageMonitor to react to changes in a package's App Lock
enabled state. This ensures that the service's internal list of App Lock
enabled packages is always updated.
- If a package has App Lock enabled, it is added to a list of tracked
  packages, and if it should be locked, listeners are notified
  immediately.
- If App Lock is disabled for a package, or if a package is removed
  entirely, the package is removed from the tracked list and listeners
  are notified that the package is no longer locked.

When the services is created, a list of all packages with App Lock
enabled is created for each user to initialize the service.

Finally, this change ensures that when the device locks, all unlocked
packages are locked immediately, and all listeners are notified of this.

Bug: 436379692
Flag: android.security.app_lock_apis
Flag: android.security.app_lock_core
Test: atest AppLockLocalService
Change-Id: I9a1921a7fda1024cac92a1cddcade9d7d46fa00f