Expose the sender of sync calls to the BinderMonitor
The BinderMonitor thread now passes the pid of the sending process to
the Java layer, as well as the pid of the receiving process.  This
allows the Java layer to implement policies against the sender, if
desired.

Tested on a build with a recent 6.12 kernel.

Flag: android.os.binder_netlink_enabled
Bug: 332610283
Test: boot Pixel with mainline kernel and check logcat
Test: atest
 * binderLibTest
 * FrameworksServicesTests:com.android.server.am
Change-Id: If3962baa5a2d9ea8204b2c6808f71af24a56053c