Convert aapt2 to protobuf lite
aapt2 is being used on the device, where protobuf lite is strongly
preferred.  The only use of a protobuf full runtime feature is
DebugString() used in ApkInfo_test.cpp.

Compile the protos with lite, add a few missing transitive includes
that get exposed by the generated lite proto headers, and replace
comparing the DebugString in the test with comparing the binary
proto.  This requires generating the proto with
CodedOutputStream.SetSerializationDeterministic(true).

Bug: 453911870
Flag: EXEMPT HOST_MODE
Test: aapt2_tests
Change-Id: Ib991d80e2188d3668c5b8fa0718911d22e21e9b9