[PiP on Desktop] Add missing RunOnTransitStart usages
DesktopTasksController#addDeskActivationChanges and
addMoveToFullscreenChanges return a RunOnTransitStart callback that
should be invoked when the transition is started.

This CL adds the missing usages for those callbacks in
DesktopPipTransitionController. Both methods are called to add wct
changes when expanding PiP.

Also modify RunOnTransitStart to be a functional interface so that we
can use it in Java classes.

Bug: 452728787
Test: atest DesktopTasksControllerTest
DesktopPipTransitionControllerTest PipSchedulerTest
Flag: EXEMPT BUGFIX

Change-Id: If0f8547a8bdc8086d5cd6ebd178f2974a84c111e