Fix predictive back system animation stuck
This CL prevents mActivityTaskManager#startPredictiveBackAnimation being
called twice for the same back navigation. This happened when a back
gesture was cancelled and immediately restarted. In that case the
post commit animation is interrupted and the leashes are animated by
the gesture again.

Recent enhancements around the predictive_back_delay_wm_transition
introduced this regression.

Bug: 454242769
Flag: EXEMPT BUGFIX
Test: BackAnimationControllerTest
Change-Id: Ic8b474ed981350ae5179d8d231cc290c6e4ae076