Refactor session priority handling.
Instead of going from lower to higher priority sessions overriding
light states, the implementation will go from higher to lower
priority only adding if the slot for light is available.

This is needed to support scenarios when lights from a session form
a block that plays together (a light effect for example), and should
only play when all the lights are available for the session to use.

Bug: 455718316
Fixes: 455718316
Flag: EXEMPT PURE_REFACTOR
Test: presubmit and unit testing.
Change-Id: I7a802384c923619b0b3773820b9a1e64c32da061