[Roll forward with fix] Use BackupManagerMonitor from KeyValueBackupTask constructor if it's not null.
This is a roll forward of ag/36474235, with a fix to prevent b/454280205. In that test, the monitor was null, but the test was still expecting a reporter. I changed getReporter() so that it still creates a reporter in the cases where the monitor is null.

Change-Id: I152f95a2b18aba9fc761d376a3e7a0eeb6405e6e
Bug: 454280205
Bug: 443738546
Bug: 380051012
Bug: 450630210
Flag: com.android.server.backup.enable_kv_backup_logs_from_transport_with_proper_flow_id
Tested: KeyValueBackupTest, GtsBackupTestCases, CtsBackupTestCases and CtsBackupHostTestCases