Test for launching fullscreen app in bubble
Test that first launches the test app in fullscreen. Then navigates back
to the home screen and launches the same app in a bubble.

Tests that an app running in fullscreen can relaunch in bubble.

One test for regular app and one for an app launching via a trampoline
app.

Bug: 439498605
Flag: EXEMPT TEST_ONLY
Test: atest WMShellExplicitFlickerTestsBubbles:RelaunchFullscreenToBubbleTest
Test: atest WMShellExplicitFlickerTestsBubbles:RelaunchFullscreenToBubbleTrampolineTest
Change-Id: Ib6064be715bf00e6a4c096aa9f6f6de80ac72c65