Assure rotate-and-crop mode is available before triggering camera compat.
Not all devices support all rotate-and-crop modes. When display rotation is sandboxed to something different from the real build-in display screen, this changes the result of a correct camera preview rotation calculation (see https://developer.android.com/media/camera/camera2/camera-preview), therefore rotate-and-crop needs to be applied to accomodate for this shift. If the required rotate and crop is not available, simulate requested orientation treatment cannot be applied.

Flag: EXEMPT BUGFIX
Fixes: 455822268
Test: atest WmTests:AppCompatCameraSimReqOrientationPolicyTests
Test: atest WmTests:CameraStateMonitorTests
TEst: manual, tried on a device that doesn't support rotate-and-crop
Change-Id: Ib4fdb52092c8c0116540dd604e643bc0933ef53b