Get rid of windowing mode early checks for requestFullscreenMode
The original implementation uses Activity#mCurrentConfig. That is set
when the activity is attached to a window and when the configuration has
updates that have public differences. Windowing mode changes aren't
guaranteed to be updated.

The resource configurations are updated more often than
Activity#mCurrentConfig, but still aren't guaranteed to have the latest
windowing mode.

There is a reliable client-side source of windowing mode at
ActivityClientRecord#mLastReportedWindowingMode. It isn't accessible
from Activity class directly, so we need to move it to Activity if we'd
like to use that. Considering that the purpose of this early check is to
save an IPC, and that illegal fullscreen mode requests aren't very
often, the loss from always sending an IPC and rely on server side
checks isn't unacceptable.

We need to fix an issue in the server-side check to achieve equivalence.

Also allow WM OWNERS to own FullscreenRequestHandler.java.

Bug: 450598129
Test: atest --iteration 10 FreeformWindowingModeTests
Flag: EXEMPT BUGFIX
Change-Id: I429a315ec613970bd30ecba1a0418609bebbf458