Remove deprecated dumpline can add private profile
Tradefed had previously been relying on parsing dumpsys for the line
"can add private profile:" to ascertain whether a private profile is
supported. Tradfed has since been updated to use a proper, robust
mechanism instead. Consequently, this is no longer required in the bug
report, and the code dedicated to it can be cleaned up and follow the
general pattern of all other profile types.

The information is still available in the bug report via other means.
The purpose of this line had been specifically for parsing.

Bug: 413464199
Bug: 446712293

Test: tapas tradefed-all && m && tools/tradefederation/core/tradefed.sh run singleCommand host -n --class com.android.tradefed.device.TestDeviceTest

Test: atest FrameworksMockingServicesTests:com.android.server.pm.UserManagerServiceMockedTest

Test: atest CtsProviderUiTestCases:android.providerui.cts.DocumentsUICrossProfileTest   then make sure that the [run-on-private-profile] is using the new tradefed check and not doing the dumpsys

Flag: android.multiuser.consistent_max_users
Change-Id: Iadbf9ccb5246eb96a1268cb55e96dd4e9a59141f