Use unique names for Mouse Keys virtual device
This change updates the name used for the virtual device created for
mouse keys to be unique. The creation and deletion of virtual devices is
asynchronous, and when the `MouseKeysInterceptor` class is rapidly
constructed/destructed it is possible for a virtual device to be created
before the previous virtual device is cleaned up. While the input stack
will handle the case of stale virtual devices, we should add this further
safeguard from the a11y stack.

Bug: 454467940
Test: Manually on device
Test: atest frameworks/base/services/tests/servicestests/src/com/android/server/accessibility/AccessibilityInputFilterInputTest.kt \
        frameworks/base/services/tests/servicestests/src/com/android/server/accessibility/AccessibilityInputFilterTest.java \
        frameworks/base/services/tests/servicestests/src/com/android/server/accessibility/AccessibilityManagerServiceTest.java \
        frameworks/base/services/tests/servicestests/src/com/android/server/accessibility/MouseKeysInterceptorTest.kt
Flag: EXEMPT BUGFIX

Change-Id: Ie88ebf70ce68398fc994bda2ecc82130a96e0f8e