Reset inputShown in clearCurMethodAndSessions
Currently we call clearCurMethodAndSessions in two cases: when the IME
is disconnected (e.g. the process is killed), and when we unbind the
IME. For the first case we always reset inputShown in the
VisibilityStateComputer. However for the second case we only started
resetting it in [1], behind a feature flag.

The IME is no longer running in both cases, so it cannot be visible. We
should always reset the inputShown inside the function call to avoid
duplication.

  [1]: Ib04023b5391b07fb829329ae1683efd63f98d358

Flag: EXEMPT PURE_REFACTOR
Bug: 281029564
Test: atest CtsInputMethodTestCases
Change-Id: I057ddc82eaa49f2e40dd238b24e6d988d7974b2a