Add some test coverage for the session death handler.
Unfortunately this basically tests this by directly calling the handler,
which is simulating the termination but not actually doing the same
thing.

However, getting a proper test that actually triggers the binder
linkToDeath triggers is much more challenging, and this at least lets us
test the handler code even if it's being triggered artificially.

This also includes a bit of refactoring to make the session
implementation into an explicit death handler itself rather than using a
lambda. This makes it much easier to reuse and call directly.

Bug: 454613215
Test: mmma, atest LskfResetTests
Flag: android.app.lskfreset.flags.enable_lskf_reset_manager
Change-Id: Ib20e86c7ec7ee2d3954aea4e02029e1a1ee1ed69