Listen for strongAuthChanges in the alternateBouncerViewBinder
To directly trigger showing the primary bouncer if primary
authentication (pin/pattern/password) is now required.
Currently, AlternateBouncerViewBinder uses the same
AlternateBouncerViewModel for each AlternateBouncerView
it adds to the window. There may be race conditions with
activating the view model; therefore, listen for the
strong auth change in the ViewBinder directly instead
of through the ViewModel.

When SceneContainer is enabled, a new
AlternateBouncerViewModel is used for each AlternateBouncer,
so it's fine to continue to rely on the ExclusiveActivatable
which is already activated in AlternateBouncer's rememberViewModel
method call (no change to this logic in this CL).

Flag: EXEMPT BUGFIX
Fixes: 448687473
Test: atest AlternateBouncerViewModelTest
Test: on a device with strong fingerprint and strong
  face: enroll fingerprint; enroll face; tap on a
  notification to show the alternate bouncer;
  fail face auth until lockout on the alternate
  bouncer; observe primary bouncer shows
Change-Id: Id5f093341826816c78e4b445ac242d8d412df0a5