[PiP on CD] Fix stashed state not updating for external displays
ActivityTaskManagerService#onPictureInPictureUiStateChanged uses
mRootWindowContainer.getDefaultTaskDisplayArea(), which means that when
PiP is not on the default display, the PiP stashed state is never sent
to the correct task on the external display. This led to a bug where
YouTube, for instance, is not informed of the stashed state in order to
pause playback on the external display.

This CL adds a displayId parameter to the
onPictureInPictureUiStateChanged method, so that if the displayId is not
the DEFAULT_DISPLAY, we can find the right tda and task to dispatch the
PiP UI state change to. The displayId will use the current displayId
stored in PipDisplayLayoutState.

Bug: 449857733
Test: Manual - stash YT PiP on external display and verify playback is
paused
Flag:com.android.window.flags.enable_pip_ui_state_change_callback_for_connected_displays

Change-Id: I05f2cced2fa5e03fbac7f66a0a94afbc15ef42e9