PackageManager: Add FEATURE_USB_DEVICE
This adds FEATURE_USB_DEVICE (android.hardware.usb.device) which is to
be set to indicate support for USB device mode. Historically,
FEATURE_USB_ACCESSORY has been abused for this use case, but this causes
issues when testing for devices which do not support the Android Open
Accessory (AOA) mode, but do have device controllers.

Certain classes of devices are not required to have
FEATURE_USB_ACCESSORY, but still provide the gadget functionality
associated with USB Device mode. This FEATURE_USB_DEVICE allows apps
which require the gadget functionality but are not reliant on
FEATURE_USB_ACCESSORY to choose to not have their app installed on
devices which do not support the required USB Device mode functionality.

Bug: 457480832
Bug: 444013436
Bug: 456622996
Test: TH
Flag: android.hardware.usb.flags.enable_feature_usb_device
Change-Id: I2feb025241e77bd9a7af269db2553a64c0917193
Signed-off-by: Neill Kapron <nkapron@google.com>