Attempt to reduce flakiness of SizeCompatTests
Though it is normal to have a WAKE transition when adding a task
to an empty display, it is unrelated to the purpose of this test
class. So simply set to not sleeping to skip the transition, that
avoids potential noise of queued transition, which can also speed
up the tests.

Fix: 454223779
Flag: EXEMPT TEST_ONLY
Test: SizeCompatTests
Change-Id: I838e32787363504a160ec2995cf85a8ff0e3546a