Use single transaction to handle default multiple animations
This is similar as the SurfaceAnimationRunner in WM core. When
a frame handles multiple animations, it only needs to put the
operations to a transaction rather than applying different
instances individually.

This can reduce the amount of binder transaction to surfaceflinger
when there are multiple default animations, e.g. regular activity
and task switch, which contains an open and a close animation.

Bug: 456665218
Flag: com.android.window.flags.default_animator_single_transaction
Test: DefaultTransitionHandlerTest
Change-Id: Ide22349833300032d98137bd3b954a11ca42ef4f