Fix bitmap leak in UserIconDrawable
When a UserIconDrawable based on a icon bitmap is baked, the source
bitmap is recycled. However, the underlying bitmap data may not be freed
because the BitmapShader can retain a native shared pointer to the
native bitmap object. Unfortunately there isn't a recycle method on
BitmapShader, so we just have to drop the java reference and wait for
the shader to be GC'ed.

Bug: 456812413
Test: Manually verify heap dumps on desktop device
Flag: EXEMPT BUGFIX
Change-Id: I3ff8b44af5e11845799f1a8e5fc21c001c122023