Don't allow multiple instances of app bubbles in the overflow
There's no visual distinction between app bubbles in the overflow so
don't allow duplicates -- favor the one with category launcher on the
intent, or the most recent. Might re-vist this if there's a way to
help users differentiate.

This changes adds a step to overflowing to determine if an app bubble
should be added to the overflow or not.

Flag: com.android.wm.shell.remove_app_bubble_overflow_dupes
Bug: 415103355
Test: atest BubbleDataTest
Test: manual - add a gmail bubble, click compose
             - dismiss both bubbles
             - add a different bubble, check the overflow
             => only 1 gmail bubble is added
             - add that bubble back
             => verify it's gmail and not compose
Change-Id: If83fe08eea32c0d91a83a43fc28ddba96eb29333