Introduce compat change to ignore BAL callbacks
Introduce a new app compatibility change, CALLBACK_IS_NOOP, to
conditionally ignore the Background Activity Launch (BAL) callback.
This supports the long-term goal of preventing background activity
launches by default for apps targeting newer SDKs.

The change is enabled for applications targeting Tiramisu (Android T)
or later. The logic in BackgroundLaunchProcessController is updated to
check this flag. The BAL callback (via the originating token) is now
ignored if *both* the global flag (Flags.balIgnoreCallback()) is
true *and* this new per-app compat change is enabled for the package.

Bug: 447255745
Flag: EXEMPT BUGFIX
Test: atest BackgroundActivityLaunchTest NotificationManagerTest
Change-Id: I04b68c594669df9820ee538f9b4a71abaaa56935