[flexiglass] Disable in-window animation if we are waking directly to Gone.
This happens in both wake and unlock and keyguard disabled cases, and the spec is to play the reveal animation while the content underneath does not move. Currently, we're preparing the launcher for unlock and then relying on a failsafe to make the icons visible at the end of the transition.

Bug: 452894970
Test: unlock from AOD w/ UDFPS wake and unlock
Test: disable keyguard and wake from AOD
Test: disable AOD+keyguard and wake from DOZING
Flag: com.android.systemui.scene_container
Change-Id: If44f001fe008ba8737a8a8355149eee0baa31f5d