Fix PasswordBouncer not properly showing IME
The flexiglass password bouncer was not properly showing the IME
since the window was not focusable during the transition from
keyguard to bouncer. This left the IME in a broken state which
required opening of the bouncer or refocus to fix. This change updates
the window flags as it was on legacy bouncer: whenever the
authentication method is changed to password, update the window
flags. This also allows to remove the temporary fix which was
refocusing the password field.

Flag: com.android.systemui.scene_container
Bug: 445347474
Bug: 451546441
Test: manual unlocking on akuta after sleep
atest: NotificationShadeWindowModelTest
Change-Id: Ibfc4f5115ad468996492aab75bf3cd5635768e57