Add test for RouteListingPreference update refreshing devices
Add a new unit test to `InfoMediaManagerTest` to verify that the device list is refreshed when the `RouteListingPreference` is updated.

This test simulates a preference update from `MediaRouter2` and confirms that the `onDeviceListAdded` callback is triggered. This ensures that the device refresh mechanism is correctly invoked, guaranteeing that the UI will reflect the latest device ordering specified by the preference.



This CL is generated by AI, please check the comments for more details.

Test: ATP tests passed http://go/forrest-run/L32100030033657785
Bug: 431235865 (Umbrella)
Flag: TEST_ONLY



Change-Id: I8dafd228b6c357778d297ed97ace92c833665359