Add test for reparentTasks' reparentTopOnly parameter
Add a test case to verify the `reparentTopOnly` parameter in the `reparentTasks` method of `WindowContainerTransaction`. This ensures correct behavior following a bug fix where this parameter was previously ignored.

The new test confirms that when `reparentTopOnly` is true, only the top-most child task is reparented. It also verifies that when the flag is false, all child tasks are moved as expected.

This CL is generated by AI, please check the comments for more details.

Test: atest services/tests/wmtests/src/com/android/server/wm/WindowContainerTransactionTests.java
Test: ATP tests passed http://go/forrest-run/L76400030034099845
Bug: 431235865 (Umbrella)
Flag: TEST_ONLY

Change-Id: Ie8d649ec380a52de013a98722c1903077f2e0eaa