audio: Use Optional for RouteSelectionStack
Migrate RouteSelectionStack to use Optional for clarity, since
preferences within the stack may now be null.

Retaining clients without an explicit will be used for external requests
to de-select SCO in the short-term, as well as enabling fusing with the
mode stack, and potentially keeping track of ordering independent of
calls to clear/disable devices.

Bug: 454733095
Flag: EXEMPT PURE_REFACTOR
Test: compiles
Change-Id: I6a6a69641a0dbed6928774a5fb98d3de61f2d84c