[Ravenwood] Support SettingsProvider
Instead of supporting the real SettingsProvider, providing an
implementation that always returns nothing is good enough for most unit
testing needs. For tests that require specific values of Settings,
TestableContext is design specifically for this need.

To make android.provider.Settings and TestableContext working as-is, we
need to enable several APIs in the content provider/resolver subsystem.

- ContentProviderNative: basically a manually written binder shim
  similar to what AIDL will generate, should always be enabled.
- ContentProviderClient: a simple wrapper around IContentProvider
  without too much additional logic.
- Enable all APIs in ContentProvider/Resolver which is just a utility
  method over its internal Context.

To ensure the implementation works correctly, add
RavenwoodSettingsProviderTest to core tests.

Bug: 292141694
Bug: 457840588
Flag: EXEMPT HOST_MODE
Test: f/b/r/scripts/run-ravenwood-tests.sh
Change-Id: I4039f71edb95662d48f6bf6564a343ba9db8a426