[PiP] Don't show menu on hover if PiP is stashed.
Currently, clicking PiP with a mouse while it's stashed does not unstash
the PiP (but tapping with a touch does) because the code to unstash
doesn't get called if the menu is visible. Hovering over the PiP window
with a mouse always shows the menu, which is why clicking afterwards
doesn't unstash the PiP.

This CL updates the showMenu code on hover enter to not show the menu if
PiP is stashed.

Bug: 450069394
Test: Manual - hover over stashed PiP and verify menu doesn't show.
Click on the PiP and verify it unstashes.
Test: atest PipTouchHandlerTest
Flag: EXEMPT BUGFIX

Change-Id: I5b8237c01d3bae6af657cb6855243ef069893e08