[AppOp] Add AppOp for Continuing Across Devices
Add a new AppOpInfo, `CONTINUE_ACROSS_DEVICES`, which specifies if Handoff is allowed for the given application. By default, this will be allowed. A future user-facing setting will be added to let the user revoke this for a package.

More information on the Handoff feature can be found here: go/handoff-design

Flag: android.companion.task_continuity
Bug: 400970610
Test: Built

Change-Id: I3f07d19d773a47d5da682d5167231864b1b638b5