psc: Refactor PSC APIs to use internal record classes
This CL refactors public APIs in `ProcessStateController` and its
`ProcessLruUpdater` interface to use the `ProcessRecordInternal`
abstraction instead of the concrete `ProcessRecord` class.

This change is part of an ongoing effort to modularize the process
state management logic. By depending on an interface, we decouple the
`ProcessStateController` from the implementation details of
`ProcessRecord`, which improves testability and overall code structure.

Bug: 457589118
Test: m services.core
Test: atest MockingOomAdjusterTests ActivityManagerServiceTest
Test: atest FrameworksServicesTestsRavenwood_ProcessStateController
Flag: EXEMPT PURE_REFACTOR

Change-Id: I89f018e02e6b9d6785233561a03807789d8494bd