Fix ContextHintWithSignature require hints to be marshallable
Now we allow the hint to provide it's own data for Parcelization that
is marshalled. By default this is just the bundle data in the hint,
but notification hints override this to include the data we know that
we care about.

This data is never read from the parcel; it is only ever used when
signing the hint.

Bug: 434644900
Test: atest PersonalContextTests
Flag: android.service.personalcontext.enable_personal_context_service
Change-Id: I4c8caaddcbe02ef9b516daaad9bec0216a6a6964