Add new fields to BridgedNotificationMetadata
This change adds originChannelId and localAppUid to the
BridgedNotificationMetadata class. These will later be used in
subsequent cls to check if the origin package of the notification
being mirror should not be posted to the local device.

originDeviceName is also added which will be used for displaying
information in the notification guts.

Change-Id: I65ea33de8a210ee144e68180bd60b2380cd13eec
Bug: 438827600
Test: atest core/tests/coretests/src/android/app/NotificationTest.java
Flag: android.app.bridged_notifications