Remove redundant display id in drag end
The displayId param is not necessary for this method. The taskInfo
contains the display id to which the task belongs before the drag
happens, and the motion event contains the display id of the final
position. Having an extra display id is not only sending extra non
necessary information, but also made the code hard to understand and not
intuitive.

This change removes the display id and replace it with the display id in
the motion event to clearly indicate that the ending display should be
used.

Test: DesktopTasksControllerTest
Test: DesktopModeTouchEventListenerTest
Test: DesktopModeWindowDecorViewModelTests
Bug: 435317310
Flag: EXEMPT PURE_REFACTOR
Change-Id: I1e628774e8ab2f6429c9c23cbd82ac34e4d59244