Add SurfaceView API for disabling accessibility hierarchy embedding.
This change introduces new methods to `SurfaceView` to programmatically
control accessibility hierarchy embedding from a child
`SurfaceControlViewHost.SurfacePackage`.

By default, `SurfaceView` automatically exposes the embedded hierarchy's
accessibility tree. This behavior is not always desirable, such as when
the `SurfaceView` is used for a purely visual preview and its content is
either decorative or replicated elsewhere in the parent's accessibility
tree.

Flag: EXEMPT BUGFIX
Bug: 445870603
Test: atest CtsSurfaceControlTestsStaging

Change-Id: Ied4c3874048fd1cdb613bd4ff9817f87c79516ee