Check IME client focus before checking user switch
Currently in IMMS we check whether we have a pending user switch task
before checking if the given IME client can actually startInput or not
(i.e. whether it is on the right display, and part of the current IME
Layering Target). This causes an issue when a particular race condition
is met when launching an existing app A (from some user profile) in
split screen with a new app B (from a different user profile). When B is
focused, it calls startInputOrWindowGainedFocus, which will see it is
not the mCurrentImeUserId, but a profile of it, and thus schedule an
user switch. If A sends another startInputOrWindowGainedFocus call
before this user switch finishes, it will schedule another user switch
onto itself, cancelling the one for B. However, A is no longer part of
the current IME Layering Target, and so it shouldn't be allowed to
request switching to its user.

When launching in split screen, A receives a configuratiion change,
which will call InputMethodManager#restartInput. This only passes if the
input was not finished already, and in this case it will call
startInputOrWindowGainedFocus. The race condition is between B calling
startInput, scheduling its user switch, and A calling restartInput
after this, but before the switch is completed.

This moves the pending switch check after the IME Client check,
essentially reverting most of the changes from [1].

  [1]: I733a8fc813d751c8529a15d5800476af4a61c3dd

Flag: EXEMPT BUGFIX
Test: atest MultiUserMockImeTest#testHidingKeyboardInSplitScreenWithCrossProfileAppsCanShowKeyboardAgain
Bug: 451594103
Change-Id: I71627eb07990a6b49e3359e44fdf51884be7355a