Fix MessageQueueTest for concurrent MQ
This change updates concurrent MQ's peekLastMessageForTest() to use
Message.compareMessages(), which considers insert order as well as
timestamps.

This change also updates testPeekLastMessageWithEqualWhen to more
reliably fail, and changes usages of sendEmptyMessageDelayed with
sendEmptyMessageAtTime.

peekLastMessageForTest() returns the last message in the MQ, which is
the most recently submitted message with the greatest timestamp. The
implementation of this method for concurrent MQ is to use
findOrRemoveMessages() (which iterates through the MQ's internal data
structures) with a comparator that stores the most recently seen message
with a timestamp greater than or equal to the previously-found "last"
message.

This is racy because findOrRemoveMessages() processes the stack first
(where messages are initially submitted) before the queues (where
messages are moved to from the stack). If findOrRemoveMessages() reaches
the queues after messages have started moving from the stack, we'll get
an incorrect "last" message because the queue's messages were actually
submitted first.

Test: FrameworksCoreTestsRavenwood
      :MessageQueueTest#testPeekLastMessage
      :MessageQueueTest#testPeekLastMessageWithEqualWhen
Bug: 452987782
Flag: EXEMPT bugfix
Change-Id: I7c0f2ab11d365a51b2363a7e08ed0512ab6a1da8